Return results in a stable order
